backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
neo4j==5.19.0
python-dotenv==1.0.1
google-generativeai==0.8.0
pydantic==2.9.0
orjson==3.10.0
httpx==0.27.0
